vector-engine/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.6.3
requests==2.31.0
packaging